Convert memory allocation functions to use size_t.
This makes memory allocation less susceptible to integer overflow
errors.

Zone allocator can't return more than 2 GiB per call, but cumulative
size of all allocations can exceed signed 32-bit integer range if
USE_MULTI_SEGMENT is enabled. For this reason convert zone statistics to
use size_t as well.

Hunk allocator continues to use signed 32-bit integers internally
because total hunk size is limited to 2 GiB.

Check upper bound of com_zoneMegs/com_hunkMegs cvars to prevent integer
overflows.

+++ code/server/sv_bot.c
@@ -290,7 +290,7 @@ static void BotImport_BSPModelMinsMaxsOrigin(int modelnum, vec3_t angles, vec3_t
 BotImport_GetMemory
 ==================
 */
-static void *BotImport_GetMemory(int size) {
+static void *BotImport_GetMemory(size_t size) {
 	void *ptr;
 
 	ptr = Z_TagMalloc( size, TAG_BOTLIB );
@@ -311,7 +311,7 @@ static void BotImport_FreeMemory(void *ptr) {
 BotImport_HunkAlloc
 =================
 */
-static void *BotImport_HunkAlloc( int size ) {
+static void *BotImport_HunkAlloc( size_t size ) {
 	if( Hunk_CheckMark() ) {
 		Com_Error( ERR_DROP, "%s(): Alloc with marks already set", __func__ );
 	}
